% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bodyView}
\alias{bodyView}
\title{Body View}
\usage{
bodyView()
}
\value{
Body's component
}
\description{
Handles the body component.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
    bodyView()
}

}
\author{
Fazendaaa
}
