% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbConnection}
\alias{dbConnection}
\title{Database Connection}
\usage{
dbConnection()
}
\value{
mongo's connection object
}
\description{
Creates a new MongoDB's connection by reading the following environment
variables:

\itemize{
\item DB_URI
\item DB_USERNAME
\item DB_PASSWORD
\item DB_HOST
\item DB_PORT
\item DB_NAME
}
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
    dbConnection()
}

}
\author{
Fazendaaa
}
\keyword{internal}
