% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbFindAll}
\alias{dbFindAll}
\title{Database Find All}
\usage{
dbFindAll()
}
\value{
The read content
}
\description{
Goes through the whole database fetching all of its content.
}
\details{
This is allowed in our case due to only using on database, collection and
content saved to it.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
   dbFindAll()
}

}
\author{
Fazendaaa
}
\keyword{internal}
