% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{dbInsert}
\alias{dbInsert}
\title{Database Insert}
\usage{
dbInsert(values)
}
\arguments{
\item{values}{\itemize{
\item content to add to database.
}}
}
\value{
The saved data
}
\description{
Save content to Mongo's database.
}
\details{
This is allowed in our case due to only using on database, collection and
content saved to it.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
    dbInsert(list(foo = 'bar'))
}

}
\author{
Fazendaaa
}
\keyword{internal}
