% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{dbView}
\alias{dbView}
\title{Database View}
\usage{
dbView(input, output, session)
}
\arguments{
\item{input}{\itemize{
\item Object containing page defined values
}}

\item{output}{\itemize{
\item Object to write response to it
}}

\item{session}{\itemize{
\item Session related information
}}
}
\value{
Database's view
}
\description{
Handles the database view page.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
   dbView()
}

}
\author{
Fazendaaa
}
