% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{eratosthenesSieve}
\alias{eratosthenesSieve}
\title{Eratosthenes Sieve}
\usage{
eratosthenesSieve(limit)
}
\arguments{
\item{limit}{\itemize{
\item upper limit to check for
}}
}
\value{
the primes numbers up to limit
}
\description{
Eratosthenes Sieve
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
    eratosthenesSieve(1)
    eratosthenesSieve(10)
    eratosthenesSieve(100)
}

}
\author{
Fazendaaa
}
\keyword{internal}
