% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hasEratosthenes}
\alias{hasEratosthenes}
\title{Has Eratosthenes}
\usage{
hasEratosthenes(set)
}
\arguments{
\item{set}{\itemize{
\item provide values to check for primes in it
}}
}
\value{
whether or not primes are presented
}
\description{
Goes throught the given set of values and check whether or not contains any
prime number in it.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
   hasEratosthenes(1:10)
   hasEratosthenes(1:100)
   hasEratosthenes(1:1000)
}

}
\author{
Fazendaaa
}
\keyword{internal}
