% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{headerView}
\alias{headerView}
\title{Header View}
\usage{
headerView()
}
\value{
Header's component
}
\description{
Handles the header component.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
   headerView()
}

}
\author{
Fazendaaa
}
