% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{histogramController}
\alias{histogramController}
\title{Histogram Controller}
\usage{
histogramController(breaks)
}
\arguments{
\item{breaks}{\itemize{
\item number of breaks
}}
}
\value{
Number of breaks
}
\description{
Adds desired the desired number of breaks to Mongos's database.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
histogramController(20)
histogramController(200)
histogramController(2000)

}
\author{
Fazendaaa
}
\keyword{internal}
