% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{histogramModel}
\alias{histogramModel}
\title{Histogram Model}
\usage{
histogramModel(breaks)
}
\arguments{
\item{breaks}{\itemize{
\item number of breaks
}}
}
\value{
The breaks number
}
\description{
As the previous implementations had no modification to the number of breaks
provided by the user, this function acts just as a "identity function". But
if any logic related to changing this value is to be implemented, it would
be done right here.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
histogramModel(1)
histogramModel(10)
histogramModel(100)

}
\author{
Fazendaaa
}
\keyword{internal}
