% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{histogramView}
\alias{histogramView}
\title{Histogram View}
\usage{
histogramView(input, output, session)
}
\arguments{
\item{input}{\itemize{
\item Object containing page defined values
}}

\item{output}{\itemize{
\item Object to write response to it
}}

\item{session}{\itemize{
\item Session related information
}}
}
\value{
Histogram's page
}
\description{
Handles the histogram view page.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
   histogramView()
}

}
\author{
Fazendaaa
}
