% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{sidebarView}
\alias{sidebarView}
\title{Sidebar View}
\usage{
sidebarView()
}
\value{
Sidebars's component
}
\description{
Handles the sidebar component.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
\dontrun{
   sidebarView()
}

}
\author{
Fazendaaa
}
