% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{dbController}
\alias{dbController}
\title{Database Controller}
\usage{
dbController()
}
\value{
The saved content
}
\description{
Fecthes the saved content from Mongo's Database if \code{HAS_DB} is set.
}
\details{
If \code{HAS_DB} is not configured, it will just provide message saying that.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
dbController()

}
\author{
Fazendaaa
}
\keyword{internal}
