% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{histogramController}
\alias{histogramController}
\title{Histogram Controller}
\usage{
histogramController(breaks)
}
\arguments{
\item{breaks}{\itemize{
\item number of breaks
}}
}
\value{
Number of breaks
}
\description{
Adds desired the desired number of breaks to Mongos's dabase if \code{HAS_DB} is
configured then return the histogram containing the density.
}
\details{
If \code{HAS_DB} is not configured, it will just return the graphic. And if
\code{HAS_API} is configured, it will use it to configure the breaks values.
}
\section{Maintainers}{

Fazendaaa
}

\examples{
histogramController(20)
histogramController(200)
histogramController(2000)

}
\author{
Fazendaaa
}
\keyword{internal}
